*: refactor cache
Move all cache maintenance into the cache code, instead supplying a ReaderAt
interface to a single cached Get() interface.  In the httpreader case, the
supplied ReaderAt uses an http client performing range GET requests.

Some shuffling occurs, like moving the length/size query into the cache
constructor which can now fail up front if the length cannot be known, rather
than in the httpreader.Size() call.

Capitalized things as if the cache is in its own package, which will happen
next, it's truly generalized and useful outside of here.

This refactor also implements more robust caching; the gets don't need to be
aligned to cached block boundaries.  Regardless of what offset a Get() is done
at, if the data is in the cache it will be utilized.

Locking needed to be reworked to use a simple mutex, since even pure cached
reads modify the LRU linked list, my LRU changes were broken since that stuff
occurred in an RLock.

When a cache miss occurs, the ReaderAt.Read() may be called concurrently, if
the cache caller is performing Get()s on the same cache concurrently.  The
cache data structures are protected with a mutex to permit concurrent usage, and
the lock is released for the duration of the ReaderAt.Read().